fbt, scripts: provisioning improvements and fixes
